+++ setup.cfg
@@ -1,15 +1,15 @@
 [metadata]
-name = stactools-ephemeral
-version = attr: stactools.ephemeral.__version__
+name = stactools-worldpop
+version = attr: stactools.worldpop.__version__
 description = PROVIDE DESCRIPTION HERE
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = stac-utils
 author_email = stac@radiant.earth
-url = https://github.com/stactools-packages/ephemeral
+url = https://github.com/stactools-packages/worldpop
 project_urls =
-    Documentation = https://stactools-ephemeral.readthedocs.io/en/latest/
-    Issues = https://github.com/stactools-packages/ephemeral/issues
+    Documentation = https://stactools-worldpop.readthedocs.io/en/latest/
+    Issues = https://github.com/stactools-packages/worldpop/issues
 keywords =
     stactools
     pystac
