+++ tests/test_module.py
@@ -1,8 +1,8 @@
 import unittest
 
-import stactools.ephemeral
+import stactools.worldpop
 
 
 class TestModule(unittest.TestCase):
     def test_version(self):
-        self.assertIsNotNone(stactools.ephemeral.__version__)
+        self.assertIsNotNone(stactools.worldpop.__version__)
